3 - Modificació: canviar el tipus de variable dels dos operadors (int a float) i el tipus de variable que retornen els métodes (int a float).

+++ src/pkg2daw_m08_uf4_recu/ArxiuAVersionar.java
@@ -11,13 +11,13 @@ package pkg2daw_m08_uf4_recu;
  */
 public class ArxiuAVersionar {
     
-    int op1, op2;
+    float op1, op2;
     
-    public int Suma(){
+    public float Suma(){
         return op1 + op2;
     }
     
-    public int Resta(){
+    public float Resta(){
         return op1 - op2;
     }
 }
